closes #14 - Tratamento de Exceções Produto

+++ src/main/java/com/luciana/desafio/services/ProdutoService.java
@@ -4,11 +4,15 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.luciana.desafio.entities.Produto;
 import com.luciana.desafio.repositories.ProdutoRepository;
-import com.luciana.desafio.services.exceptions.ObjectNotFoundException;
+import com.luciana.desafio.services.exceptions.DatabaseException;
+import com.luciana.desafio.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProdutoService {
@@ -24,7 +28,7 @@ public class ProdutoService {
 	
 	public Produto findById(Integer id) {
 		Optional<Produto> obj = repository.findById(id);
-		return obj.orElseThrow(() -> new ObjectNotFoundException("Objeto nao encontrado! Id: " + id + ", Tipo: " + Produto.class.getName()));
+		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
 	
 	
@@ -36,18 +40,32 @@ public class ProdutoService {
 	
 	
 	
-	// Para deletar produto
+	// Para deletar produto - ARRUMAR CÓDIGO PARA INATIVAR PRODUTO ASSICIADO A VENDA
 	public void deletar(Integer id) {
-		repository.deleteById(id);
+		
+		try {
+			if (repository.existsById(id)) {
+				repository.deleteById(id);			
+			} else {				
+				throw new ResourceNotFoundException(id);			
+			}		
+		} catch (DataIntegrityViolationException e) {			
+			throw new DatabaseException(e.getMessage());		
+		}		
 	}
 	
 	
 	
 	// Para atualizar um produto
 	public Produto atualizar(Integer id, Produto obj) {
-		Produto entity = repository.getReferenceById(id);
-		atualizarDados(entity, obj);
-		return repository.save(entity);		
+		try {
+			Produto entity = repository.getReferenceById(id);
+			atualizarDados(entity, obj);
+			return repository.save(entity);	
+		}
+		catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException(id);	
+		}
 	}
 
 	private void atualizarDados(Produto entity, Produto obj) {
